refactor: read the parent process through internal/platform
The Win32 calls move out of the alias package into platform, which
owns the per-OS conventions. The alias package keeps the mapping from
an image name to a shell, which a table test covers on every OS. The
sandbox doctor test joins doctor_test.go.
