--- internal/alias/shells_other.go
@@ -1,5 +1,0 @@
-//go:build !windows
-
-package alias
-
-func parentShellName() string { return "" }
