+++ internal/platform/process_unix.go
@@ -1,0 +1,7 @@
+//go:build unix
+
+package platform
+
+import "errors"
+
+func parentProcessName() (string, error) { return "", errors.ErrUnsupported }
